# Conflicts:
#	Antenna/Tracker.de-DE.resx
#	ChangeLog.txt
#	Controls/LogAnalyzer.cs
#	Controls/MavlinkCheckBox.cs
#	Controls/MyDataGridView.cs
#	Controls/Video.resx
#	Controls/Waypoints/Spline2.cs
#	CurrentState.cs
#	ElevationProfile.cs
#	ExtLibs/AltitudeAngelWings/Service/AltitudeAngelService.cs
#	ExtLibs/Arduino/Arduino.csproj
#	ExtLibs/AviFile/_UpgradeReport_Files/UpgradeReport.css
#	ExtLibs/AviFile/_UpgradeReport_Files/UpgradeReport.xslt
#	ExtLibs/AviFile/packages.config
#	ExtLibs/BSE.Windows.Forms/Properties/Resources.Designer.cs
#	ExtLibs/BaseClasses/packages.config
#	ExtLibs/Comms/CommsBase.cs
#	ExtLibs/Comms/CommsNTRIP.cs
#	ExtLibs/Comms/CommsSerialPipe.cs
#	ExtLibs/Comms/CommsTCPSerial.cs
#	ExtLibs/Comms/CommsUDPSerialConnect.cs
#	ExtLibs/Comms/CommsUdpSerial.cs
#	ExtLibs/Comms/SerialTest.cs
#	ExtLibs/Comms/packages.config
#	ExtLibs/Controls/HUD.cs
#	ExtLibs/Controls/MissionPlanner.Controls.csproj
#	ExtLibs/Controls/MyButton.cs
#	ExtLibs/Core/CommonObjects/IdNameColorItem.cs
#	ExtLibs/Core/FileHandling/LoadSave.cs
#	ExtLibs/Core/Utils/ColorUtils.cs
#	ExtLibs/Core/Xml/XmlSerializer.cs
#	ExtLibs/GMap.NET.Core/GMap.NET.Internals/PureImage.cs
#	ExtLibs/GMap.NET.Core/GMap.NET.Internals/Stuff.cs
#	ExtLibs/GMap.NET.Core/GMap.NET.MapProviders/GMapProvider.cs
#	ExtLibs/GMap.NET.WindowsForms/GMap.NET.WindowsForms/Markers/GMarkerGoogle.cs
#	ExtLibs/GStreamerHud/GStreamerHud.csproj
#	ExtLibs/GlobalAssemblyInfo.cs
#	ExtLibs/Grid/MissionPlanner.Grid.csproj
#	ExtLibs/LibVLC.NET/Properties/AssemblyInfo.cs
#	ExtLibs/ManagedNativeWifi.Simple/App.config
#	ExtLibs/ManagedNativeWifi.Simple/ManagedNativeWifi.Simple.csproj
#	ExtLibs/ManagedNativeWifi.Simple/packages.config
#	ExtLibs/Mavlink/MAVLinkParamList.cs
#	ExtLibs/Mavlink/MavlinkParse.cs
#	ExtLibs/Mavlink/message_definitions/ardupilotmega.xml
#	ExtLibs/Mavlink/message_definitions/common.xml
#	ExtLibs/MetaDataExtractorCSharp240d/packages.config
#	ExtLibs/MissionPlanner.Gridv2/GridPluginv2.cs
#	ExtLibs/MissionPlanner.Gridv2/GridUIv2.cs
#	ExtLibs/SharpAdbClient/AdbClient.cs
#	ExtLibs/SharpAdbClient/SharpAdbClient.csproj
#	ExtLibs/SharpAdbClient/packages.config
#	ExtLibs/SimpleExample/SimpleExample.csproj
#	ExtLibs/SimpleExample/simpleexample.cs
#	ExtLibs/SimpleGrid/GridUI.pt.resx
#	ExtLibs/SimpleGrid/SimpleGridUI.cs
#	ExtLibs/SimpleGrid/SimpleGridUI.resx
#	ExtLibs/Utilities/CleanDrivers.cs
#	ExtLibs/Utilities/DisplayView.cs
#	ExtLibs/Utilities/Math.cs
#	ExtLibs/ZedGraph/ZedGraph/RollingPointPairList.cs
#	ExtLibs/px4uploader/app.config
#	ExtLibs/px4uploader/packages.config
#	ExtLibs/solo/Program.cs
#	ExtLibs/solo/Solo.cs
#	ExtLibs/tlogThumbnailHandler/tlogThumbnailHandler.csproj
#	FollowMe.de-DE.resx
#	GCSViews/ConfigurationView/ConfigAP_Limits.de-DE.resx
#	GCSViews/ConfigurationView/ConfigAP_Limits.fr.resx
#	GCSViews/ConfigurationView/ConfigAccelerometerCalibration.de-DE.resx
#	GCSViews/ConfigurationView/ConfigArducopter.de-DE.resx
#	GCSViews/ConfigurationView/ConfigArduplane.de-DE.resx
#	GCSViews/ConfigurationView/ConfigArdurover.de-DE.resx
#	GCSViews/ConfigurationView/ConfigArdurover.pt.resx
#	GCSViews/ConfigurationView/ConfigAteryx.de-DE.resx
#	GCSViews/ConfigurationView/ConfigAteryxSensors.de-DE.resx
#	GCSViews/ConfigurationView/ConfigBatteryMonitoring.de-DE.resx
#	GCSViews/ConfigurationView/ConfigBatteryMonitoring2.de-DE.resx
#	GCSViews/ConfigurationView/ConfigESCCalibration.cs
#	GCSViews/ConfigurationView/ConfigFailSafe.de-DE.resx
#	GCSViews/ConfigurationView/ConfigFlightModes.de-DE.resx
#	GCSViews/ConfigurationView/ConfigHWBT.cs
#	GCSViews/ConfigurationView/ConfigHWCompass.Designer.cs
#	GCSViews/ConfigurationView/ConfigHWCompass.resx
#	GCSViews/ConfigurationView/ConfigMotorTest.Designer.cs
#	GCSViews/ConfigurationView/ConfigMotorTest.cs
#	GCSViews/ConfigurationView/ConfigMotorTest.resx
#	GCSViews/ConfigurationView/ConfigMount.cs
#	GCSViews/ConfigurationView/ConfigMount.de-DE.resx
#	GCSViews/ConfigurationView/ConfigPlanner.cs
#	GCSViews/ConfigurationView/ConfigPlanner.de-DE.resx
#	GCSViews/ConfigurationView/ConfigPlanner.id-ID.resx
#	GCSViews/ConfigurationView/ConfigPlanner.resx
#	GCSViews/ConfigurationView/ConfigRadioInput.de-DE.resx
#	GCSViews/ConfigurationView/ConfigRawParams.de-DE.resx
#	GCSViews/ConfigurationView/ConfigRawParamsTree.de-DE.resx
#	GCSViews/ConfigurationView/ConfigTradHeli.Designer.cs
#	GCSViews/ConfigurationView/ConfigTradHeli.cs
#	GCSViews/ConfigurationView/ConfigTradHeli.de-DE.resx
#	GCSViews/FlightData.Designer.cs
#	GCSViews/FlightData.cs
#	GCSViews/FlightData.de-DE.resx
#	GCSViews/FlightData.resx
#	GCSViews/FlightPlanner.Designer.cs
#	GCSViews/FlightPlanner.cs
#	GCSViews/FlightPlanner.resx
#	GCSViews/Help.de-DE.resx
#	GCSViews/Terminal.de-DE.resx
#	GeoRef/Georefimage.Designer.cs
#	GeoRef/Georefimage.de-DE.resx
#	GeoRef/georefimage.cs
#	GeoRef/georefimage.resx
#	Grid/Grid.cs
#	Grid/GridUI.Designer.cs
#	Grid/GridUI.cs
#	Grid/GridUI.resx
#	Joystick/JoystickSetup.Designer.cs
#	Joystick/JoystickSetup.de-DE.resx
#	Joystick/JoystickSetup.resx
#	Log/DFLogScript.cs
#	Log/LogBrowse.cs
#	Log/LogBrowse.resx
#	Log/LogIndex.cs
#	Log/MatLab.cs
#	LogAnalyzer/py2exe/tests/TestVCC.py
#	MagCalib.cs
#	MainV2.Designer.cs
#	MainV2.cs
#	MainV2.resx
#	Mavlink/MAVLinkInterface.cs
#	Mavlink/MAVState.cs
#	MissionPlanner.csproj
#	OpenGLtest2.cs
#	Program.cs
#	Properties/AssemblyInfo.cs
#	Properties/Resources.Designer.cs
#	RAW_Sensor.de-DE.resx
#	README.md
#	SerialInjectGPS.Designer.cs
#	SerialInjectGPS.cs
#	SerialInjectGPS.pt.resx
#	SerialInjectGPS.resx
#	SerialOutputNMEA.cs
#	SikRadio/app.config
#	Strings.Designer.cs
#	Updater/Properties/Resources.Designer.cs
#	Updater/Properties/Settings.Designer.cs
#	Updater/Properties/Settings.settings
#	Updater/Updater.csproj
#	Updater/app.config
#	Utilities/AP_Terrain.cs
#	Utilities/Crypto.cs
#	Utilities/ExpireType.cs
#	Utilities/GStreamer.cs
#	Utilities/GimbalPoint.cs
#	Utilities/LogAnalyzer.cs
#	Utilities/Settings.cs
#	Utilities/ThemeManager.cs
#	Utilities/UDPVideoShim.cs
#	Utilities/nmea.cs
#	Utilities/rtcm3.cs
#	Utilities/ubx_m8p.cs
#	Wizard/3ConnectAP.cs
#	Wizard/6CompassCalib.cs
#	app.config
#	appxmanifest.xml
#	convert.pl
#	m3u/networklink.kml
#	mavcmd.xml
#	packages.config
#	packages/repositories.config
#	resedit/resedit.csproj
#	temp.Designer.cs
#	temp.cs
#	wix/wix.csproj

+++ ExtLibs/Utilities/GeoTiff.cs
@@ -16,7 +16,7 @@ namespace MissionPlanner.Utilities
         private static readonly ILog log =
             LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
 
-        private static Dictionary<string, float[,]> cache = new Dictionary<string, float[,]>();
+        private static Dictionary<string, short[,]> cache = new Dictionary<string, short[,]>();
 
         public static List<geotiffdata> index = new List<geotiffdata>();
 
@@ -151,7 +151,7 @@ namespace MissionPlanner.Utilities
                     // add to cache
                     if (!cache.ContainsKey(geotiffdata.FileName) && geotiffdata.cacheable)
                     {
-                        float[,] altdata = new float[geotiffdata.height, geotiffdata.width];
+                        short[,] altdata = new short[geotiffdata.height, geotiffdata.width];
 
                         using (Tiff tiff = Tiff.Open(geotiffdata.FileName, "r"))
                         {
@@ -169,7 +169,7 @@ namespace MissionPlanner.Utilities
                                     }
                                     else if (geotiffdata.bits == 32)
                                     {
-                                        altdata[row, col] = (float) BitConverter.ToSingle(scanline, col*4);
+                                        altdata[row, col] = (short) BitConverter.ToSingle(scanline, col*4);
                                     }
                                 }
                             }
